decorate.py: Add script
The script performs a set of test preparation tasks that are usually
done by the STDCI system. Those tasks include:

1. Source cloning
2. Upstream source pulling
3. Providing the `extra_sources` file
4. Mirror injection into `*.yumrepos` files

The script is meant to be run in an init container when the whole CI
process runs in a pod and the user had set the `decorate` option to
`True` in the STDCI DSL. That is the reason behind its name.

Change-Id: I8146fa5fc6289ae2e3a8cf142316b4af1fb646b7
Signed-off-by: Barak Korren <bkorren@redhat.com>