gate: cleanup gate resources.
Added a playbook that deletes repositories objects deployed by the gate
via the deploy_to_container.yaml playbook.
There will be a dedicated jenkins cleanup job starts with 'cleanup-'
for example 'cleanup-gated-ovirt-master'

I also added functionallity to have 2 jobs in a single group under JJB:
- deploy jobs
- cleanup jobs

In the future, we can use this group to add change queues repositories
to be added here.

The cleanup policy is:
1) Latest resources are never cleaned up
2) The newest resource which is X days older or more won't be cleaned up
either. where X is the day defined in the environment variable.

I've also seperated the email notify function into a seperated library
to be shared to reduce duplicated code.

Change-Id: If13b14136058d98519e97f16007ed61c91dbc171
Signed-off-by: Ehud Yonasi <eyonasi@redhat.com>