pusher: Consider dest branch when checking if push is needed
There is a situation where the same change should be pushed to two
different branches (for example right after a project branch to a new
version, but still didn't move the development to the new branch).

This change will allow it, by considering the destination branch when
checking if a change already been submitted.

Change-Id: I6b5c3bb5949b9ccd827ccc986ef496e6d9e894d2
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2849
Signed-off-by: gbenhaim <galbh2@gmail.com>