Enable STDCI tools container to run playbooks & oc
Install `ansible` and `oc` in the stdci-tools container so playbooks
that access OpenShift could be run from it as well as general OpenShift
workloads.

The container is configured with environment variables that
automatically select the `main` inventory (Also added in this patch), so
that the `ansible-playbook` commands one needs to run in it can be
shorter.

The container default command is also changed to allow it to be used as
a sidecar container in a Jenkins slave POD.

Finally added some symlinks to make the container build process
compatible with the new locations for the STDCI tools and libs.

Change-Id: I6b6145b6e3c1852a47e165aa90b7cd0b5b8a6f7c
Signed-off-by: Barak Korren <bkorren@redhat.com>