Refactor: Move python tools to their own directory
It became an issue to have a Python module called `scripts` because
other packages seem to use that name. Taking the opportunity to move
things around to more descriptive directory names.

This patch moves all the Python command-line tools to their own
`stdci_tools` directory.

Note: leaving a symlink in `scripts` to `usrc_local.py` to enable
systems to transition to the new configuration.

Change-Id: I476ba2e2e7821bdc8b6d6184551e46cb1f6653be
Signed-off-by: Barak Korren <bkorren@redhat.com>