Refactor: Move STDCI Python libs to their own dir
It became an issue to have a Python module called `scripts` because
other packages seem to use that name. Taking the opportunity to move
things around to more descriptive directory names.

This patch moves all the STDCI Python libraries to the `stdci_libs`
directory.

Note: Some tricks were needed to ensure that object that were Pickled
when the libraries were in the old location, could be read by the
libraries from the new location. This was further complicated by
differences between how cPickle works in Python2 and Python3.

Change-Id: Iddaf45025df974dde914b48fa990b231bd62c978
Signed-off-by: Barak Korren <bkorren@redhat.com>