dockerfile_utils: Inject SHA sums
Allow users to inject the SHA sum of a parent image, instead of
the NVR tag. This is usefull for users who prefer to use SHA
sum for specifying the version of the parent images.
It's usefull as well for parent images that doesn't contain the
"name", "version" and "release" labels, since for those a static
tag can't be calculated.

Change-Id: Ida425e3cdcaddea74cfca8fa043f7c3d02dc424f
Signed-off-by: gbenhaim <galbh2@gmail.com>