run_code: added a script to run on node.
with dynamic provisioning of nodes, we can use pre-configured node to
run the code we want.

This change introduces couple of changes because we run without mock and
it's added features:

1) In order to use the ci_toolbox like in mock, I've mounted it to the
same location: /var/lib/ci_toolbox and added it to the $PATH.

2) To achieve the environment variables like in mock, I've had to dump
the yaml from the secrets_resolver and to parse it in bash and source
the script. After that, I need to enable the env vars into the user
script, therefore Instead of calling it, i am souring the script so it
will run in the parent process(run_code).

3) I've modified the check-patch file to run the mock checks in upstream, and
added a test to run privileged container in podman in psi.

4) Added cases to run the code only in PSI jenkins in stdci_runner.

Change-Id: Ib2b83c28ea883ec49970ff5d45b67d87fecfd756
Signed-off-by: Ehud Yonasi <eyonasi@redhat.com>