dockerfile-utils: Better handling for nested repos
Instead of replacing the entire repository name in the
floating reference, replace just the tag with the static
tag or sha sum (depends on the user's request).

This change removes the complexity of normalizing nested
repositories which are not supported in the "rh-osbs" namespace,
but supported in the stage and prod registries.

Add tests to cover image updates from the prod registry (getting
updates from tage is the same, the url just changes from
registry.redhat.io to registry.stage.redhat.io).

Change-Id: I366a71900ffb20cb741d1669b0f9cb08fd851344
Signed-off-by: gbenhaim <galbh2@gmail.com>