usrc: Add post update hook
The hook will allow users to run commands after usrc updates
the upstream-sources config, and before commiting it.

The stdout of the hook will be added to the commit message, so
it would be possible to describe the changes the hook made to
to the upstream sources config in the commit message.

The logic that checks if a new commit should be created was changed,
since now it's possible that the hook will change the upstream-sources
config, thus a new commit should be created.

This is usefull, for example, when a user want to have a custom
update algorithem.

Jira-Url: https://issues.redhat.com/browse/KNIECO-2330
Change-Id: I3f2ad3dd46fe597e906197f47c6680418f6fb4cd
Signed-off-by: gbenhaim <galbh2@gmail.com>