Add automerge flag for updaters
Now that updater main's API enables updaters to apply commits
(CNV-8795), it is sometimes required to include the 'automerge' in
commit headers as well. This is required for version_updater and for
more updaters to come.

To enable automerge, updaters can now add the '--automerge' CLI
parameter flag, and pass it to updater_main.
Respectivey, this parameter should be set in project.yml for updaters
that support it.

The automerge flag is only applicable for updaters that use the
framework to submit commits. Thus, for e.g., usrc_updater will not be
affected by this change.

Jira-URL: https://issues.redhat.com/browse/CNV-9075
Signed-off-by: Guy Hartuv <ghartuv@redhat.com>
Change-Id: I833d5cfc3528c17e59b658dce59772c10ce6c277