Beaker integration.
This patch aims to use beaker to reserve hosts and register them
to jenkins master.

It has several features:
1) Added a beaker xml for rhel8 host configuration, which can be used
for OST setup in the future. It also register the beaker host to the jenkins
master.

2) Added a stage to call for the beaker system in order to reserve a host from
the system. with that host reserved we know which host will be registering to jenkins, thus
making the labeling specific to that host inside the job thread.

3) Added a post actions to return the beaker host to the system and deleting it from jenkins.

Change-Id: Ibb7df58b20bf24eb737529b4511bea3be94aafad
Signed-off-by: Ehud Yonasi <eyonasi@redhat.com>