SubScript called from Main script cause main script to fail
The main script calls  a sub script. the sub script is exiting the
script using exit and cause the main script to stop running.
In order to fix this issue we will call the sub script in subshell.
Subshells are separate processes, so exit exits only them.
When we exit the main script without continuing it cause the
CI to show that the script actually pass althou it failed.

Change-Id: I73a7cfac6fe52b4525d42e90af560026864b2e40
Signed-off-by: Galit <grosenth@redhat.com>