mock: c9s: move config_opts['root'] to the top
without this change mock_runner is inserting something like:

  config_opts['root'] = 'centos-stream-9-x86_64-89bbfb8bed34dbd898a2bd36790571e0'

at the top of the generated mock config but it will be overridden by
the following definition at the end of the file.

This caused the same chroot cache to be used across the different builds
without installing the required dependencies.

Change-Id: Id4ed363e569967ce126cabee681219444e2faad4
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>